backend/pyproject.toml
[project]
name = "here-inference-api"
version = "0.1.0"
requires-python = "==3.12.10"
description = "HERE Inference API"
readme = "README.md"
dependencies = [
    # FastAPI and related
    "fastapi~=0.112.2",
    "uvicorn~=0.34.1",
    "starlette~=0.37.2",
    "h11~=0.14.0",
    "click~=8.1.7",
    "anyio~=4.2.0",
    "sniffio~=1.3.1",
    "python-multipart~=0.0.9",

    # Pydantic and typing
    "pydantic[email,timezone]~=2.7.1",
    "pydantic-settings~=2.2.1",
    "pydantic-core~=2.18.2",
    "typing-extensions~=4.12.2",
    "annotated-types~=0.7.0",
    "python-dotenv~=1.0.1",
    "email-validator~=2.1.1",
    "idna~=3.7",

    # Database and ORM
    "sqlalchemy~=2.0.40",
    "alembic~=1.15.2",
    "asyncpg~=0.30.0",
    "greenlet~=3.2.0",

    # Security and Auth
    "passlib[bcrypt]~=1.7.4",
    "python-jose[cryptography]~=3.3.0",
    "ecdsa~=0.18.0",
    "six~=1.16.0",
]

[project.optional-dependencies]
dev = [
    "pyright~=1.1.369",
    "pre-commit~=3.3.2",
    "ruff~=0.11.6",
]

[tool.ruff]
target-version = "py312"
line-length = 88
exclude = [
    "**/node_modules",
    "**/__pycache__",
    ".git",
    ".mypy_cache",
    "app/tests",
]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "TCH", "TID"]
ignore = []
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pyright]
include = ["app"]
exclude = ["**/node_modules", "**/__pycache__", "app/tests"]
pythonVersion = "3.12.10"
reportMissingImports = true
reportMissingTypeStubs = false
reportGeneralTypeIssues = "warning"
reportOptionalMemberAccess = "warning"
reportOptionalSubscript = "warning"
reportPrivateImportUsage = "warning"
reportUnboundVariable = "warning"
reportUnusedImport = "error"
reportUnusedVariable = "error"
venvPath = "."
venv = ".venv"
